ml/requirements.txt
librosa==0.10.2.post1
soundfile==0.12.1
pyyaml==6.0.2
tqdm==4.66.5
scikit-learn==1.5.2
numpy
pandas
matplotlib
torch
scipy
seaborn
